Pushed update
So this is no longer ArmorModelPredicate, because it doesn't change the model. It is now ArmorRenderLib because it is a small bit of reusable code that extends the FabricAPI ArmorRenderer to allow the dynamic changing of textures and colors of any armor item. Still in beta because while I could technically add support for predicates at this point (such as item tags), I have no need yet so I'm choosing to reduce bloat.